+++ web/src/views/product/Data.vue
@@ -93,15 +93,17 @@ export default {
       });
     },
     submit(no_product) {
-      axios.delete(`${process.env.VUE_APP_SERVICE_URL}/products/` + no_product, { headers: authHeader() })
-          .then(response => {
-            if(response.data.code === 200) {
-              alert(response.data.status)
-              this.fetchData()
-            }
-          }).catch(error => {
-        console.log(error.response.data.status)
-      })
+      if(confirm("Are you sure to delete this product?")) {
+        axios.delete(`${process.env.VUE_APP_SERVICE_URL}/products/` + no_product, { headers: authHeader() })
+            .then(response => {
+              if(response.data.code === 200) {
+                alert(response.data.status)
+                this.fetchData()
+              }
+            }).catch(error => {
+          console.log(error.response.data.status)
+        })
+      }
     }
   }
 }
