+++ web/src/views/product/Detail.vue
@@ -156,17 +156,19 @@ export default {
       })
     },
     submit(no_product) {
-      axios.delete(`${process.env.VUE_APP_SERVICE_URL}/products/` + no_product, { headers: authHeader() })
-          .then(response => {
-            if(response.data.code === 200) {
-              alert(response.data.status)
-              this.$router.push({
-                name: 'product'
-              })
-            }
-          }).catch(error => {
-            console.log(error.response.data.status)
-          })
+      if(confirm("Are you sure to delete this product?")) {
+        axios.delete(`${process.env.VUE_APP_SERVICE_URL}/products/` + no_product, { headers: authHeader() })
+            .then(response => {
+              if(response.data.code === 200) {
+                alert(response.data.status)
+                this.$router.push({
+                  name: 'product'
+                })
+              }
+            }).catch(error => {
+          console.log(error.response.data.status)
+        })
+      }
     }
   }
 }
