+++ web/src/views/user/Data.vue
@@ -92,15 +92,17 @@ export default {
       });
     },
     submit(id) {
-      axios.delete(`${process.env.VUE_APP_SERVICE_URL}/users/` + id, { headers: authHeader() })
-          .then(response => {
-            if(response.data.code === 200) {
-              alert(response.data.status)
-              this.fetchData()
-            }
-          }).catch(error => {
-        console.log(error.response.data.status)
-      })
+      if(confirm("Are you sure to delete this data?")) {
+        axios.delete(`${process.env.VUE_APP_SERVICE_URL}/users/` + id, { headers: authHeader() })
+            .then(response => {
+              if(response.data.code === 200) {
+                alert(response.data.status)
+                this.fetchData()
+              }
+            }).catch(error => {
+          console.log(error.response.data.status)
+        })
+      }
     }
   }
 }
